creating separated form component
