+++ package.json
@@ -21,7 +21,6 @@
     "react-is": "^16.8.0",
     "recoil": "^0.1.3",
     "styled-components": "^5.0.0",
-    "sweetalert2": "^11.0.16",
     "yup": "^0.32.9"
   },
   "devDependencies": {
